Major Update Data Pull
I adjusted the organization of the functions and variables to the inherit global location. and tried to account for our access token expiring, but it was not successful. I tested my creating a MockResponse Class and assigning it to the hardware_response var so that I could manually set a 400 error to test my theory which did not pan out, but may have been to Also Energy revoking access, i had to reset our credentials this morning. And I am now simply waiting for a natural 400 error for further testing,.