Date Validation Rule Added
Updated window frames and add the Date Validation Rule created for the WO to Logbook Tool. Probably should've added it to python tools package and imported it, but I might have to change it or make it program specifc, not sure yet. This is the first time I have used a validation rule